Solution for Cablescan since latest Ziggo changes
Ziggo is currently switching channels to from MPEG2 to AVC (MPEG4). This means they are currently using the HD channelid list for SD channels already changed to AVC and the SD channelid list has dummy channels. When in the scan the AVC channel was scanned before the MPEG2 dummy channel the AVC channel was "overwritten" in the userbouquets by the dummy MPEG2 channel. This commit solves that one a channelid is used it will not be overwitten by a SD (MPEG2) channel